requirements.txt
streamlit==1.42.2
pymongo==1.42.2
python-dotenv==1.0.1